requirements.txt
python-dotenv==1.1.0
mcp[cli]==1.9.4
openai-agents==0.0.19